+++ src/components/Parallax/index.ts
@@ -1,2 +1,3 @@
 export { default as OrigamParallaxElement } from './OrigamParallaxElement.vue'
+export { default as OrigamParallaxLayer } from './OrigamParallaxLayer.vue'
 export { default as OrigamParallax } from './OrigamParallax.vue'