+++ src/interfaces/index.ts
@@ -199,6 +199,7 @@ export * from './Pagination/pagination.interface'
 
 export * from './Parallax/parallax.interface'
 export * from './Parallax/parallax-element.interface'
+export * from './Parallax/parallax-layer.interface'
 
 export * from './PasswordField/password-field.interface'
 export * from './PasswordField/password-requirement.interface'