+++ src/types/Parallax/parallax-easing.type.ts
@@ -1,0 +1,3 @@
+import { PARALLAX_EASING } from '../../enums'
+
+export type TParallaxEasing = `${PARALLAX_EASING}`